+++ src/shared/event/remote_event_registry.rs
@@ -35,21 +35,21 @@ impl RemoteEventRegistry {
         self.client_triggers.push(trigger);
     }
 
-    pub(crate) fn iter_server_events_mut(&mut self) -> impl Iterator<Item = &mut ServerEvent> {
-        self.server_events.iter_mut().chain(
-            self.server_triggers
-                .iter_mut()
-                .map(|trigger| trigger.event_mut()),
-        )
+    pub(super) fn iter_server_events_mut(&mut self) -> impl Iterator<Item = &mut ServerEvent> {
+        self.server_events.iter_mut()
     }
 
-    pub(crate) fn iter_server_events(&self) -> impl Iterator<Item = &ServerEvent> {
+    pub(super) fn iter_server_triggers_mut(&mut self) -> impl Iterator<Item = &mut ServerTrigger> {
+        self.server_triggers.iter_mut()
+    }
+
+    pub(crate) fn iter_all_server(&self) -> impl Iterator<Item = &ServerEvent> {
         self.server_events
             .iter()
             .chain(self.server_triggers.iter().map(|trigger| trigger.event()))
     }
 
-    pub(crate) fn iter_client_events(&self) -> impl Iterator<Item = &ClientEvent> {
+    pub(crate) fn iter_all_client(&self) -> impl Iterator<Item = &ClientEvent> {
         self.client_events
             .iter()
             .chain(self.client_triggers.iter().map(|trigger| trigger.event()))
@@ -68,7 +68,7 @@ impl RemoteEventRegistry {
     /// See also [`ServerEventAppExt::add_server_event`](super::server_event::ServerEventAppExt::add_server_event)
     /// and [`ServerTriggerAppExt::add_server_trigger`](super::server_trigger::ServerTriggerAppExt::add_server_trigger).
     pub fn server_channel<E: Event>(&self) -> Option<usize> {
-        self.iter_server_events()
+        self.iter_all_server()
             .find(|event| event.type_id() == TypeId::of::<E>())
             .map(|event| event.channel_id())
     }
@@ -78,7 +78,7 @@ impl RemoteEventRegistry {
     /// See also [`ClientEventAppExt::add_client_event`](super::client_event::ClientEventAppExt::add_client_event)
     /// and [`ClientTriggerAppExt::add_client_trigger`](super::client_trigger::ClientTriggerAppExt::add_client_trigger).
     pub fn client_channel<E: Event>(&self) -> Option<usize> {
-        self.iter_client_events()
+        self.iter_all_client()
             .find(|event| event.type_id() == TypeId::of::<E>())
             .map(|event| event.channel_id())
     }